+++ src/Event/UpdateSubmissionPostEvent.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\FormalizeBundle\Event;
 
-class CreateSubmissionPostEvent extends AbstractSubmissionEvent
+class UpdateSubmissionPostEvent extends AbstractSubmissionEvent
 {
 }